refactor(LC040): split mixed-tracking analysis helpers
- separate tracking-mode analysis, context resolution, and assigned-value analysis
- keep analyzer behavior unchanged while reducing local complexity
